refactor webpack configs
