+++ README.md
@@ -9,8 +9,9 @@
 ```
 npm run dev # запуск проекта локально
 npm run build # сборка проекта
-npm run start # сборка + запуск проекта на 3000 порту
+npm run start:prod # запуск проекта на 3000 порту
 npm run lint # ESLint + Stylelint
+npm run test # Jest tests
 ```
 
 Макеты в figma: https://www.figma.com/file/Dhb1EOJFJ2U1X430XoKSTr/yandex-praktikum-messenger?node-id=0%3A1
