+++ package.json
@@ -6,11 +6,12 @@
         "dev": "webpack-dev-server --mode development",
         "start": "npm run build && node server.js",
         "start:prod": "node server.js",
-        "build": "cross-env NODE_ENV=production webpack --config ./webpack/webpack.config.prod.js",
-        "dev-build": "cross-env NODE_ENV=development webpack-dev-server --hot --config ./webpack/webpack.config.dev.js",
+        "build": "cross-env NODE_ENV=production webpack --config webpack.config.prod.js",
+        "dev-build": "cross-env NODE_ENV=development webpack-dev-server --hot --config webpack.config.dev.js",
         "lint": "eslint ./src/**/*.ts --fix && stylelint ./src/**/*.css --fix",
         "test": "jest",
-        "pre-commit": "npm run test && npm run lint"
+        "pre-commit": "npm run test && npm run lint",
+        "webpack-validate": "webpack configtest webpack.config.dev.js"
     },
     "engines": {
         "node": "18.12.1"
